+++ hybridpos/src/main/java/com/example/hybridpos/dto/ProductCreateDTO.java
@@ -1,5 +1,7 @@
 package com.example.hybridpos.dto;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
@@ -8,8 +10,8 @@ public class ProductCreateDTO {
     private String barcode;
     private String name;
 
-    private double purchasePrice;
-    private double salePrice;
+    private BigDecimal purchasePrice;
+    private BigDecimal salePrice;
 
     private int stockQuantity;
 }
